Fixed SkullRenderer not working with Forge
